minor tweaks
