+++ SRC/system_of_eqn/linearSOE/sparsePython/SparsePythonCommon.h
@@ -1,6 +1,8 @@
 #ifndef SparsePythonCommon_h
 #define SparsePythonCommon_h
 
+#include <Python.h>
+
 /**
  * Storage scheme class: acts like an enum with string conversion.
  */
@@ -92,6 +94,40 @@ struct SparsePythonWritableFlags {
 
     SparsePythonWritableFlags() = default;
     SparsePythonWritableFlags(bool v, bool r) : values(v), rhs(r) {}
+};
+
+/**
+ * RAII helper that manages reference counting for transient PyObject* instances.
+ * Acquires the GIL before calling Py_XDECREF to ensure thread safety.
+ */
+struct PyObjectHolder {
+    PyObjectHolder() = default;
+    explicit PyObjectHolder(PyObject *obj) : ptr(obj) {}
+    ~PyObjectHolder() {
+        // Acquire GIL before calling Py_XDECREF to ensure thread safety
+        // Note: We don't set ptr to nullptr here because the object is being destroyed
+        if (ptr != nullptr && Py_IsInitialized()) {
+            PyGILState_STATE gilState = PyGILState_Ensure();
+            Py_XDECREF(ptr);
+            PyGILState_Release(gilState);
+        }
+    }
+    PyObject *get() const { return ptr; }
+    PyObject *release() {
+        PyObject *tmp = ptr;
+        ptr = nullptr;
+        return tmp;
+    }
+    void reset(PyObject *obj = nullptr) {
+        if (ptr == obj) {
+            return;
+        }
+        // reset() should only be called while GIL is held
+        Py_XDECREF(ptr);
+        ptr = obj;
+    }
+  private:
+    PyObject *ptr{nullptr};
 };
 
 #endif /* SparsePythonCommon_h */
