+++ SRC/system_of_eqn/linearSOE/sparsePython/SparsePythonCompressedLinSolver.cpp
@@ -1,7 +1,6 @@
 #include "SparsePythonCompressedLinSolver.h"
 
 #include "SparsePythonCompressedLinSOE.h"
-#include "SparsePythonCommon.h"
 
 #include <Channel.h>
 #include <FEM_ObjectBroker.h>
@@ -9,33 +8,7 @@
 #include <cstring>
 #include <elementAPI.h>
 #include <Python.h>
-
-
-namespace {
-
-// RAII helper that manages reference counting for transient PyObject* instances.
-struct PyObjectHolder {
-    PyObjectHolder() = default;
-    explicit PyObjectHolder(PyObject *obj) : ptr(obj) {}
-    ~PyObjectHolder() { Py_XDECREF(ptr); }
-    PyObject *get() const { return ptr; }
-    PyObject *release() {
-        PyObject *tmp = ptr;
-        ptr = nullptr;
-        return tmp;
-    }
-    void reset(PyObject *obj = nullptr) {
-        if (ptr == obj) {
-            return;
-        }
-        Py_XDECREF(ptr);
-        ptr = obj;
-    }
-  private:
-    PyObject *ptr{nullptr};
-};
-
-} // namespace
+#include "SparsePythonCommon.h"
 
 SparsePythonCompressedLinSolver::SparsePythonCompressedLinSolver()
     : LinearSOESolver(SOLVER_TAGS_SparsePythonCompressedLinSolver),
@@ -57,17 +30,15 @@ SparsePythonCompressedLinSolver::SparsePythonCompressedLinSolver(PyObject *calla
 
 SparsePythonCompressedLinSolver::~SparsePythonCompressedLinSolver()
 {
-    if (solverObject != nullptr) {
-        if (Py_IsInitialized()) {
-            // Acquire GIL before decrementing reference count
-            // This is critical: Py_XDECREF must be called with GIL held
-            PyGILState_STATE gil_state = PyGILState_Ensure();
-            Py_XDECREF(solverObject);
-            PyGILState_Release(gil_state);
-        }
-        // If Python is not initialized, we cannot safely call Py_XDECREF
-        // The object will be cleaned up when Python interpreter shuts down
-        solverObject = nullptr;
+    // NOTE: This is generally unsafe to call Python C-API from destructors,
+    // as destructors may run after Py_Finalize(). This will be addressed later.
+    // For now, we assume the user will call ops.wipe() at the end of their script
+    // to ensure resources are released before exiting Python, making it safe to
+    // decrement refcounts in destructors.
+    if (solverObject != nullptr && Py_IsInitialized()) {
+        PyGILState_STATE gilState = PyGILState_Ensure();
+        Py_XDECREF(solverObject);
+        PyGILState_Release(gilState);
     }
 }
 
@@ -85,17 +56,24 @@ SparsePythonCompressedLinSolver::setPythonCallable(PyObject *callable, const cha
         methodName = method;
     }
 
+    // Acquire GIL before any Python API calls
+    PyGILState_STATE gilState = PyGILState_Ensure();
+
     if (callable == nullptr) {
         Py_XDECREF(solverObject);
         solverObject = nullptr;
+        PyGILState_Release(gilState);
         return 0;
     }
 
     Py_XINCREF(callable);
     Py_XDECREF(solverObject);
     solverObject = callable;
+
+    PyGILState_Release(gilState);
     return 0;
 }
+
 
 int
 SparsePythonCompressedLinSolver::solve(void)
@@ -257,6 +235,9 @@ SparsePythonCompressedLinSolver::callPythonSolver()
     }
     PyObjectHolder argsHolder(args);
 
+    // PyObject_Call must be made with GIL held. CuPy/NumPy will release the GIL
+    // internally during their native GPU/CPU computations, so we don't need to
+    // do anything special here.
     PyObjectHolder result(PyObject_Call(method.get(), argsHolder.get(), kwargs.get()));
     if (result.get() == nullptr) {
         opserr << "SparsePythonCompressedLinSolver::callPythonSolver - Python callable raised an exception\n";
